Resource sharing FIR Filter
It's base of first part FIR module! Check it via testbench to see if it works!
Mention, The output of the sum of the first part module is 34 bits, while It's assigned to a 32-bit register! Because of that, I also used a 32-bit register! Tell me if anything needs to be changed!