xilinx core_test update
in the previous core test i've set the clock 50 MHz and  checked just the few output data and they were correct. but from tenth data to the end they were wrong and i did not check::(((
so i reset the clock on 0.0441 MHz and the good news is now the more of them as i checked are correct , maybe all:)))
in this test the output is 16 bit length MSB. but i'll put the output.txt  in 34 bit length too.